GetGraphs -> ListGraphs; added GetSchema method
