+++ mongo/graph.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"log"
 	"strings"
-	"time"
 
 	"github.com/bmeg/arachne/aql"
 	"github.com/bmeg/arachne/engine/core"
@@ -17,222 +16,11 @@ import (
 	"github.com/globalsign/mgo/bson"
 )
 
-// Config describes the configuration for the mongodb driver.
-type Config struct {
-	URL                    string
-	DBName                 string
-	Username               string
-	Password               string
-	BatchSize              int
-	UseAggregationPipeline bool
-}
-
-// Mongo is the base driver that manages multiple graphs in mongo
-type Mongo struct {
-	database string
-	conf     Config
-	session  *mgo.Session
-	ts       *timestamp.Timestamp
-}
-
-// NewMongo creates a new mongo graph database interface
-func NewMongo(conf Config) (gdbi.GraphDB, error) {
-	log.Printf("Starting Mongo Driver")
-	database := strings.ToLower(conf.DBName)
-	err := aql.ValidateGraphName(database)
-	if err != nil {
-		return nil, fmt.Errorf("invalid database name: %v", err)
-	}
-
-	ts := timestamp.NewTimestamp()
-	dialinfo := &mgo.DialInfo{
-		Addrs:    []string{conf.URL},
-		Database: conf.DBName,
-		Username: conf.Username,
-		Password: conf.Password,
-		AppName:  "arachne",
-	}
-	session, err := mgo.DialWithInfo(dialinfo)
-	if err != nil {
-		return nil, err
-	}
-	session.SetSocketTimeout(1 * time.Hour)
-	session.SetSyncTimeout(1 * time.Minute)
-
-	b, _ := session.BuildInfo()
-	if !b.VersionAtLeast(3, 6) {
-		session.Close()
-		return nil, fmt.Errorf("requires mongo 3.6 or later")
-	}
-	if conf.BatchSize == 0 {
-		conf.BatchSize = 1000
-	}
-	db := &Mongo{database: database, conf: conf, session: session, ts: &ts}
-	for _, i := range db.GetGraphs() {
-		db.ts.Touch(i)
-	}
-	return db, nil
-}
-
-// Close the connection
-func (ma *Mongo) Close() error {
-	ma.session.Close()
-	ma.session = nil
-	return nil
-}
-
-// VertexCollection returns a *mgo.Collection
-func (ma *Mongo) VertexCollection(session *mgo.Session, graph string) *mgo.Collection {
-	return session.DB(ma.database).C(fmt.Sprintf("%s_vertices", graph))
-}
-
-// EdgeCollection returns a *mgo.Collection
-func (ma *Mongo) EdgeCollection(session *mgo.Session, graph string) *mgo.Collection {
-	return session.DB(ma.database).C(fmt.Sprintf("%s_edges", graph))
-}
-
-// AddGraph creates a new graph named `graph`
-func (ma *Mongo) AddGraph(graph string) error {
-	err := aql.ValidateGraphName(graph)
-	if err != nil {
-		return err
-	}
-
-	session := ma.session.Copy()
-	session.ResetIndexCache()
-	defer session.Close()
-	defer ma.ts.Touch(graph)
-
-	graphs := session.DB(ma.database).C("graphs")
-	err = graphs.Insert(bson.M{"_id": graph})
-	if err != nil {
-		return fmt.Errorf("failed to insert graph %s: %v", graph, err)
-	}
-
-	e := ma.EdgeCollection(session, graph)
-	err = e.EnsureIndex(mgo.Index{
-		Key:        []string{"$hashed:from"},
-		Unique:     false,
-		DropDups:   false,
-		Sparse:     false,
-		Background: true,
-	})
-	if err != nil {
-		return fmt.Errorf("failed create index for graph %s: %v", graph, err)
-	}
-	err = e.EnsureIndex(mgo.Index{
-		Key:        []string{"$hashed:to"},
-		Unique:     false,
-		DropDups:   false,
-		Sparse:     false,
-		Background: true,
-	})
-	if err != nil {
-		return fmt.Errorf("failed create index for graph %s: %v", graph, err)
-	}
-	err = e.EnsureIndex(mgo.Index{
-		Key:        []string{"$hashed:label"},
-		Unique:     false,
-		DropDups:   false,
-		Sparse:     false,
-		Background: true,
-	})
-	if err != nil {
-		return fmt.Errorf("failed create index for graph %s: %v", graph, err)
-	}
-
-	v := ma.VertexCollection(session, graph)
-	err = v.EnsureIndex(mgo.Index{
-		Key:        []string{"$hashed:label"},
-		Unique:     false,
-		DropDups:   false,
-		Sparse:     false,
-		Background: true,
-	})
-	if err != nil {
-		return fmt.Errorf("failed create index for graph %s: %v", graph, err)
-	}
-
-	return nil
-}
-
-// DeleteGraph deletes `graph`
-func (ma *Mongo) DeleteGraph(graph string) error {
-	session := ma.session.Copy()
-	defer session.Close()
-	defer ma.ts.Touch(graph)
-
-	g := session.DB(ma.database).C("graphs")
-	v := ma.VertexCollection(session, graph)
-	e := ma.EdgeCollection(session, graph)
-
-	verr := v.DropCollection()
-	if verr != nil {
-		log.Printf("Drop vertex collection failed: %v", verr)
-	}
-	eerr := e.DropCollection()
-	if eerr != nil {
-		log.Printf("Drop edge collection failed: %v", eerr)
-	}
-	gerr := g.RemoveId(graph)
-	if gerr != nil {
-		log.Printf("Remove graph id failed: %v", gerr)
-	}
-
-	if verr != nil || eerr != nil || gerr != nil {
-		return fmt.Errorf("failed to delete graph: %s; %s; %s", verr, eerr, gerr)
-	}
-
-	return nil
-}
-
-// GetGraphs lists the graphs managed by this driver
-func (ma *Mongo) GetGraphs() []string {
-	session := ma.session.Copy()
-	defer session.Close()
-
-	out := make([]string, 0, 100)
-	g := session.DB(ma.database).C("graphs")
-
-	iter := g.Find(nil).Iter()
-	defer iter.Close()
-	if err := iter.Err(); err != nil {
-		log.Printf("Error: %s", err)
-	}
-	result := map[string]interface{}{}
-	for iter.Next(&result) {
-		out = append(out, result["_id"].(string))
-	}
-	if err := iter.Err(); err != nil {
-		log.Println("GetGraphs error:", err)
-	}
-
-	return out
-}
-
-// Graph obtains the gdbi.DBI for a particular graph
-func (ma *Mongo) Graph(graph string) (gdbi.GraphInterface, error) {
-	found := false
-	for _, gname := range ma.GetGraphs() {
-		if graph == gname {
-			found = true
-		}
-	}
-	if !found {
-		return nil, fmt.Errorf("graph '%s' was not found", graph)
-	}
-	return &Graph{
-		ar:        ma,
-		ts:        ma.ts,
-		graph:     graph,
-		batchSize: ma.conf.BatchSize,
-	}, nil
-}
-
 // Graph is the interface to a single graph
 type Graph struct {
-	ar        *Mongo
-	ts        *timestamp.Timestamp //BUG: This timestamp implementation doesn't work againt multiple mongo clients
+	ar *GraphDB
+	//BUG: This timestamp implementation doesn't work againt multiple mongo clients
+	ts        *timestamp.Timestamp
 	graph     string
 	batchSize int
 }
